:sparkles: feat(product): 상품 price int로 변경후 상품 수량 추가 및 결제 시 수량 감소 구현

+++ src/main/java/com/kuku9/goods/domain/product/dto/ProductResponse.java
@@ -10,7 +10,7 @@ public class ProductResponse {
 	private Long productId;
 	private String productName;
 	private String productDescription;
-	private Long productPrice;
+	private int productPrice;
 	private String brandName;
 
 	public ProductResponse(Product product) {
